переиспользует компонент OfferCardList на странице offer-page
