+++ src/components/app/app.tsx
@@ -15,24 +15,30 @@ type Props = {
 }
 
 export default function App({ offers }: Props): JSX.Element {
-  const [activeCity, setActiveCity] = useState<string>(DEFAULT_CITY);
+  const [activeOfferId, setActiveOfferId] = useState<string | undefined>(undefined);
+
+  const handleActiveOfferChange = (id: string | undefined) => setActiveOfferId(id);
+
+  const [activeCityName, setActiveCity] = useState<string>(DEFAULT_CITY);
 
   const handleCityClick = (city: string) => {
     setActiveCity(city);
   };
+  const filteredOffers: OfferType[] = offers.filter((offer) => offer.city.name === activeCityName);
+  const offersCount: number = filteredOffers.length;
   return (
     <Routes>
-      <Route path={RoutePath.Main} element={<MainPage onHandleCityClick={handleCityClick} activeCityName={activeCity} offers={offers} />} />
+      <Route path={RoutePath.Main} element={<MainPage activeOfferId={activeOfferId} onHandleActiveOfferChange={handleActiveOfferChange} filteredOffers={filteredOffers} offersCount={offersCount} onHandleCityClick={handleCityClick} activeCityName={activeCityName} offers={offers} />} />
       <Route path={RoutePath.Login} element={<LoginPage />} />
       <Route
         path={RoutePath.Favorites}
         element={
           <PrivatRoute loginStatus={LoginStatus.Auth}>
-            <FavoritePage offers={offers}/>
+            <FavoritePage offers={offers} />
           </PrivatRoute>
         }
       />
-      <Route path={RoutePath.Offer} element={<OfferPage offers={offers}/>} />
+      <Route path={RoutePath.Offer} element={<OfferPage filteredOffers={filteredOffers} onHandleActiveOfferChange={handleActiveOfferChange} />} />
       <Route path={RoutePath.NOT_FAUND} element={<NotFound />} />
     </Routes>
   );
